change how setting mtrr works
